refactor!: restructure codebase into app module and add new components
The project structure was refactored to encapsulate all
source code under the new `app/` module.
This aligns with modern src-layout conventions and prepares
the codebase for clean API and UI integration.

Key changes:
- Moved all existing code from `src/` into `app/`
- Added new submodules 'api' & 'ui' to support upcoming FastAPI and Streamlit interfaces
- Updated all relative and absolute import paths accordingly
- Ensured compatibility with Python tooling and unit tests via `__init__.py` files

BREAKING CHANGE:
All module paths have changed. Imports like `from autovisionai.configs import ...` must
now use `from autovisionai.app.configs import ...`.
Any external references or scripts relying on the old structure will need to be updated.
