+++ src/autovisionai/app/__init__.py
@@ -1,0 +1,6 @@
+from autovisionai.app.configs import CONFIG, PROJECT_VERSION
+from autovisionai.app.loggers import setup_app_logger
+
+__version__ = PROJECT_VERSION
+
+setup_app_logger(CONFIG.logging.app_logger)
