+++ src/autovisionai/app/configs/__init__.py
@@ -1,5 +1,12 @@
-from autovisionai.configs.config import CONFIG, CONFIG_DIR, CONFIG_FILES, PROJECT_NAME, PROJECT_ROOT, PROJECT_VERSION
-from autovisionai.configs.schema import (
+from autovisionai.app.configs.config import (
+    CONFIG,
+    CONFIG_DIR,
+    CONFIG_FILES,
+    PROJECT_NAME,
+    PROJECT_ROOT,
+    PROJECT_VERSION,
+)
+from autovisionai.app.configs.schema import (
     AppLoggerConfig,
     FileLoggerConfig,
     LRSchedulerConfig,
