+++ src/autovisionai/app/configs/local/data.yaml
