+++ src/autovisionai/app/configs/local/models.yaml
