+++ src/autovisionai/app/configs/prod/models.yaml
