+++ src/autovisionai/app/loggers/ml_logging.py
@@ -13,7 +13,7 @@ import wandb
 from PIL import Image
 from pytorch_lightning.loggers import MLFlowLogger, TensorBoardLogger, WandbLogger
 
-from autovisionai.configs import CONFIG, CONFIG_DIR, MLLoggersConfig
+from autovisionai.app.configs import CONFIG, CONFIG_DIR, MLLoggersConfig
 
 logger = logging.getLogger(__name__)
