+++ src/autovisionai/app/models/fast_scnn/fast_scnn_model.py
