+++ src/autovisionai/app/models/mask_rcnn/mask_rcnn_model.py
@@ -2,7 +2,7 @@ import torchvision
 from torchvision.models.detection.faster_rcnn import FastRCNNPredictor
 from torchvision.models.detection.mask_rcnn import MaskRCNN_ResNet50_FPN_Weights, MaskRCNNPredictor
 
-from autovisionai.configs import CONFIG
+from autovisionai.app.configs import CONFIG
 
 
 def create_model(n_classes: int, pretrained: bool = True) -> MaskRCNNPredictor:
