+++ src/autovisionai/app/utils/common.py
